updating database scripts

+++ src/main/resources/db/migrations/changes/18112019_create_table_group.sql
@@ -1,5 +1,5 @@
 create table groups
 (
-    id         serial primary key not null,
+    id         bigint primary key not null,
     identifier varchar(15)
 )
